+++ app/javascript/packs/application.js
@@ -9,10 +9,10 @@
 
 import WebpackerReact from 'webpacker-react';
 import CadmusNavbarAdminApp from 'cadmus-navbar-admin';
-import StaffPositionForm from './Forms/StaffPositionForm';
-import TeamMemberForm from './Forms/TeamMemberForm';
-import TicketPurchaseForm from './Forms/TicketPurchaseForm';
-import TicketTypeForm from './Forms/TicketTypeForm';
+import StaffPositionForm from './BuiltInForms/StaffPositionForm';
+import TeamMemberForm from './BuiltInForms/TeamMemberForm';
+import TicketPurchaseForm from './BuiltInForms/TicketPurchaseForm';
+import TicketTypeForm from './BuiltInForms/TicketTypeForm';
 
 WebpackerReact.setup({
   CadmusNavbarAdminApp,
